+++ client/utils/queryInput/addQueryToDB.js
@@ -1,4 +1,4 @@
-import db from '../db';
+import db from '../../db';
 
 const addQueryToDB = (textValue, urlToSend) => new Promise((resolve, reject) => {
   console.log('running addQueryToDB');
