+++ client/utils/queryOutputDisplay/nullResultChecker.jsx
